Mrg branch

+++ services/content/src/main/java/com/genedu/content/dto/flatResponse/FlatSchoolClassSubjectDTO.java
@@ -8,7 +8,7 @@ public record FlatSchoolClassSubjectDTO(
         String schoolClassName,
         String schoolClassDescription,
 
-        Long subjectId,
+        Integer subjectId,
         String subjectName,
         String subjectDescription
 ) {
