+++ services/content/src/main/java/com/genedu/content/service/SubjectService.java
@@ -1,20 +1,19 @@
 package com.genedu.content.service;
 
-import com.genedu.content.dto.flatResponse.FlatSchoolClassResponseDTO;
 import com.genedu.content.dto.flatResponse.FlatSchoolClassSubjectDTO;
 import com.genedu.content.dto.schoolclass.SchoolClassResponseDTO;
 import com.genedu.content.dto.subject.SubjectRequestDTO;
-import com.genedu.content.dto.subject.SubjectResponseDTO;
 import com.genedu.content.model.Subject;
 
 import java.util.List;
 
 public interface SubjectService {
     List<FlatSchoolClassSubjectDTO> getAllSubjects();
-    Subject getSubjectEntityById(Long id);
-    FlatSchoolClassSubjectDTO getSubjectById(Long id);
+    Subject getSubjectEntityById(Integer id);
+    FlatSchoolClassSubjectDTO getSubjectById(Integer id);
     SchoolClassResponseDTO getSubjectsBySchoolClassId(Integer schoolClassId);
+
     FlatSchoolClassSubjectDTO createSubject(Integer schoolClassId, SubjectRequestDTO subjectRequestDTO);
-    FlatSchoolClassSubjectDTO updateSubject(Long id, SubjectRequestDTO subjectRequestDTO);
-    void deleteSubject(Long id);
+    FlatSchoolClassSubjectDTO updateSubject(Integer id, SubjectRequestDTO subjectRequestDTO);
+    void deleteSubject(Integer id);
 }
